+++ alembic/versions/d0fdcf651c1c_add_topic_model.py
@@ -1,0 +1,40 @@
+"""Add Topic model
+
+Revision ID: d0fdcf651c1c
+Revises: dce3ec9fb4b9
+Create Date: 2024-07-30 14:37:28.848431
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'd0fdcf651c1c'
+down_revision: Union[str, None] = 'dce3ec9fb4b9'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('topics',
+    sa.Column('title', sa.String(), nullable=False),
+    sa.Column('content', sa.String(), nullable=False),
+    sa.Column('tags', sa.ARRAY(sa.String()), nullable=True),
+    sa.Column('id', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_topics_id'), 'topics', ['id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_topics_id'), table_name='topics')
+    op.drop_table('topics')
+    # ### end Alembic commands ###
