--- alembic/versions/e30fc0cc2d35_create_tables.py
@@ -1,451 +1,0 @@
-"""initial migration
-
-Revision ID: e30fc0cc2d35
-Revises: 
-Create Date: 2024-07-29 22:24:00.741552
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e30fc0cc2d35'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('contact_us',
-    sa.Column('full_name', sa.String(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('message', sa.Text(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_contact_us_id'), 'contact_us', ['id'], unique=False)
-    op.create_table('faqs',
-    sa.Column('question', sa.String(), nullable=True),
-    sa.Column('answer', sa.Text(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_faqs_id'), 'faqs', ['id'], unique=False)
-    op.create_table('newsletters',
-    sa.Column('email', sa.String(length=150), nullable=False),
-    sa.Column('title', sa.String(), nullable=True),
-    sa.Column('content', sa.Text(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_index(op.f('ix_newsletters_id'), 'newsletters', ['id'], unique=False)
-    op.create_table('organizations',
-    sa.Column('company_name', sa.String(), nullable=False),
-    sa.Column('company_email', sa.String(), nullable=True),
-    sa.Column('industry', sa.String(), nullable=True),
-    sa.Column('organization_type', sa.String(), nullable=True),
-    sa.Column('country', sa.String(), nullable=True),
-    sa.Column('state', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('lga', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('company_email'),
-    sa.UniqueConstraint('company_name')
-    )
-    op.create_index(op.f('ix_organizations_id'), 'organizations', ['id'], unique=False)
-    op.create_table('product_categories',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name')
-    )
-    op.create_index(op.f('ix_product_categories_id'), 'product_categories', ['id'], unique=False)
-    op.create_table('topics',
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('content', sa.String(), nullable=False),
-    sa.Column('tags', sa.ARRAY(sa.String()), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_topics_id'), 'topics', ['id'], unique=False)
-    op.create_table('users',
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('password', sa.String(), nullable=True),
-    sa.Column('first_name', sa.String(), nullable=True),
-    sa.Column('last_name', sa.String(), nullable=True),
-    sa.Column('is_active', sa.Boolean(), server_default=sa.text('true'), nullable=True),
-    sa.Column('is_super_admin', sa.Boolean(), server_default=sa.text('false'), nullable=True),
-    sa.Column('is_deleted', sa.Boolean(), server_default=sa.text('false'), nullable=True),
-    sa.Column('is_verified', sa.Boolean(), server_default=sa.text('false'), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_table('waitlist',
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('full_name', sa.String(), nullable=False),
-    sa.Column('joined_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_waitlist_id'), 'waitlist', ['id'], unique=False)
-    op.create_table('activity_logs',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('action', sa.String(), nullable=False),
-    sa.Column('timestamp', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_activity_logs_id'), 'activity_logs', ['id'], unique=False)
-    op.create_table('billing_plans',
-    sa.Column('organization_id', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('price', sa.Numeric(), nullable=False),
-    sa.Column('currency', sa.String(), nullable=False),
-    sa.Column('duration', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('features', sa.ARRAY(sa.String()), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['organization_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_billing_plans_id'), 'billing_plans', ['id'], unique=False)
-    op.create_table('blogs',
-    sa.Column('author_id', sa.String(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('image_url', sa.String(), nullable=True),
-    sa.Column('is_deleted', sa.Boolean(), server_default=sa.text('false'), nullable=True),
-    sa.Column('excerpt', sa.Text(), nullable=True),
-    sa.Column('tags', sa.Text(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['author_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_blogs_id'), 'blogs', ['id'], unique=False)
-    op.create_table('invitations',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('organization_id', sa.String(), nullable=False),
-    sa.Column('expires_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('is_valid', sa.Boolean(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['organization_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_invitations_id'), 'invitations', ['id'], unique=False)
-    op.create_table('jobs',
-    sa.Column('author_id', sa.String(), nullable=False),
-    sa.Column('title', sa.Text(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=False),
-    sa.Column('department', sa.String(), nullable=True),
-    sa.Column('location', sa.String(), nullable=True),
-    sa.Column('salary', sa.String(), nullable=True),
-    sa.Column('job_type', sa.String(), nullable=True),
-    sa.Column('company_name', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['author_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_jobs_id'), 'jobs', ['id'], unique=False)
-    op.create_table('messages',
-    sa.Column('message', sa.Text(), nullable=False),
-    sa.Column('phone_number', sa.String(), nullable=True),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_messages_id'), 'messages', ['id'], unique=False)
-    op.create_table('notifications',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('message', sa.Text(), nullable=False),
-    sa.Column('status', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_notifications_id'), 'notifications', ['id'], unique=False)
-    op.create_table('oauth',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('provider', sa.String(), nullable=False),
-    sa.Column('sub', sa.String(), nullable=False),
-    sa.Column('access_token', sa.String(), nullable=False),
-    sa.Column('refresh_token', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_oauth_id'), 'oauth', ['id'], unique=False)
-    op.create_table('payments',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('amount', sa.Numeric(), nullable=False),
-    sa.Column('currency', sa.String(), nullable=False),
-    sa.Column('status', sa.String(), nullable=False),
-    sa.Column('method', sa.String(), nullable=False),
-    sa.Column('transaction_id', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('transaction_id')
-    )
-    op.create_index(op.f('ix_payments_id'), 'payments', ['id'], unique=False)
-    op.create_table('products',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('price', sa.Numeric(), nullable=False),
-    sa.Column('org_id', sa.String(), nullable=False),
-    sa.Column('category_id', sa.String(), nullable=False),
-    sa.Column('quantity', sa.Integer(), nullable=True),
-    sa.Column('image_url', sa.String(), nullable=False),
-    sa.Column('status', sa.Enum('in_stock', 'out_of_stock', 'low_on_stock', name='productstatusenum'), nullable=True),
-    sa.Column('archived', sa.Boolean(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['category_id'], ['product_categories.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['org_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_products_id'), 'products', ['id'], unique=False)
-    op.create_table('profiles',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('username', sa.String(), nullable=True),
-    sa.Column('pronouns', sa.String(), nullable=True),
-    sa.Column('job_title', sa.String(), nullable=True),
-    sa.Column('department', sa.String(), nullable=True),
-    sa.Column('social', sa.Text(), nullable=True),
-    sa.Column('bio', sa.Text(), nullable=True),
-    sa.Column('phone_number', sa.String(), nullable=True),
-    sa.Column('avatar_url', sa.String(), nullable=True),
-    sa.Column('recovery_email', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_profiles_id'), 'profiles', ['id'], unique=False)
-    op.create_table('testimonials',
-    sa.Column('client_designation', sa.String(), nullable=True),
-    sa.Column('client_name', sa.String(), nullable=True),
-    sa.Column('author_id', sa.String(), nullable=False),
-    sa.Column('comments', sa.Text(), nullable=True),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('ratings', sa.Float(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['author_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_testimonials_id'), 'testimonials', ['id'], unique=False)
-    op.create_table('token_logins',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('token', sa.String(), nullable=False),
-    sa.Column('expiry_time', sa.DateTime(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_token_logins_id'), 'token_logins', ['id'], unique=False)
-    op.create_table('user_newsletter_association',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('newsletter_id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['newsletter_id'], ['newsletters.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('user_id', 'newsletter_id')
-    )
-    op.create_table('user_organization',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('organization_id', sa.String(), nullable=False),
-    sa.Column('role', sa.Enum('admin', 'user', 'guest', 'owner', name='user_org_role'), nullable=False),
-    sa.Column('status', sa.Enum('member', 'suspended', 'left', name='user_org_status'), nullable=False),
-    sa.ForeignKeyConstraint(['organization_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('user_id', 'organization_id')
-    )
-    op.create_table('blog_dislikes',
-    sa.Column('blog_id', sa.String(), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('ip_address', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['blog_id'], ['blogs.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_blog_dislikes_id'), 'blog_dislikes', ['id'], unique=False)
-    op.create_table('blog_likes',
-    sa.Column('blog_id', sa.String(), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('ip_address', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['blog_id'], ['blogs.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_blog_likes_id'), 'blog_likes', ['id'], unique=False)
-    op.create_table('comments',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('blog_id', sa.String(), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['blog_id'], ['blogs.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_comments_id'), 'comments', ['id'], unique=False)
-    op.create_table('product_variants',
-    sa.Column('size', sa.String(), nullable=False),
-    sa.Column('stock', sa.Integer(), nullable=True),
-    sa.Column('price', sa.Numeric(), nullable=True),
-    sa.Column('product_id', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['product_id'], ['products.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_product_variants_id'), 'product_variants', ['id'], unique=False)
-    op.create_table('comment_dislikes',
-    sa.Column('comment_id', sa.String(), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('ip_address', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['comment_id'], ['comments.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_comment_dislikes_id'), 'comment_dislikes', ['id'], unique=False)
-    op.create_table('comment_likes',
-    sa.Column('comment_id', sa.String(), nullable=False),
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('ip_address', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['comment_id'], ['comments.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_comment_likes_id'), 'comment_likes', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_comment_likes_id'), table_name='comment_likes')
-    op.drop_table('comment_likes')
-    op.drop_index(op.f('ix_comment_dislikes_id'), table_name='comment_dislikes')
-    op.drop_table('comment_dislikes')
-    op.drop_index(op.f('ix_product_variants_id'), table_name='product_variants')
-    op.drop_table('product_variants')
-    op.drop_index(op.f('ix_comments_id'), table_name='comments')
-    op.drop_table('comments')
-    op.drop_index(op.f('ix_blog_likes_id'), table_name='blog_likes')
-    op.drop_table('blog_likes')
-    op.drop_index(op.f('ix_blog_dislikes_id'), table_name='blog_dislikes')
-    op.drop_table('blog_dislikes')
-    op.drop_table('user_organization')
-    op.drop_table('user_newsletter_association')
-    op.drop_index(op.f('ix_token_logins_id'), table_name='token_logins')
-    op.drop_table('token_logins')
-    op.drop_index(op.f('ix_testimonials_id'), table_name='testimonials')
-    op.drop_table('testimonials')
-    op.drop_index(op.f('ix_profiles_id'), table_name='profiles')
-    op.drop_table('profiles')
-    op.drop_index(op.f('ix_products_id'), table_name='products')
-    op.drop_table('products')
-    op.drop_index(op.f('ix_payments_id'), table_name='payments')
-    op.drop_table('payments')
-    op.drop_index(op.f('ix_oauth_id'), table_name='oauth')
-    op.drop_table('oauth')
-    op.drop_index(op.f('ix_notifications_id'), table_name='notifications')
-    op.drop_table('notifications')
-    op.drop_index(op.f('ix_messages_id'), table_name='messages')
-    op.drop_table('messages')
-    op.drop_index(op.f('ix_jobs_id'), table_name='jobs')
-    op.drop_table('jobs')
-    op.drop_index(op.f('ix_invitations_id'), table_name='invitations')
-    op.drop_table('invitations')
-    op.drop_index(op.f('ix_blogs_id'), table_name='blogs')
-    op.drop_table('blogs')
-    op.drop_index(op.f('ix_billing_plans_id'), table_name='billing_plans')
-    op.drop_table('billing_plans')
-    op.drop_index(op.f('ix_activity_logs_id'), table_name='activity_logs')
-    op.drop_table('activity_logs')
-    op.drop_index(op.f('ix_waitlist_id'), table_name='waitlist')
-    op.drop_table('waitlist')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_topics_id'), table_name='topics')
-    op.drop_table('topics')
-    op.drop_index(op.f('ix_product_categories_id'), table_name='product_categories')
-    op.drop_table('product_categories')
-    op.drop_index(op.f('ix_organizations_id'), table_name='organizations')
-    op.drop_table('organizations')
-    op.drop_index(op.f('ix_newsletters_id'), table_name='newsletters')
-    op.drop_table('newsletters')
-    op.drop_index(op.f('ix_faqs_id'), table_name='faqs')
-    op.drop_table('faqs')
-    op.drop_index(op.f('ix_contact_us_id'), table_name='contact_us')
-    op.drop_table('contact_us')
-    # ### end Alembic commands ###
